# openapi/audit.yaml
openapi: 3.0.3
info:
  title: Audit API
  version: 0.1.0
paths:
  /audit/provider-errors:
    get:
      summary: List provider error events
      operationId: listProviderErrors
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: provider
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: org_id
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /audit/dlq:
    get:
      summary: List dead-letter queue entries
      operationId: listDLQ
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: queue
          schema:
            type: string
        - in: query
          name: org_id
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        request_id:
          type: string